Factor AtomicBorrow and EntityRef/Ref into separate modules
